requirements.txt
Flask==2.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.1.0
Werkzeug==2.1.2
click==8.1.3
colorama==0.4.4
colorthief==0.2.1
itsdangerous==2.1.2
numpy==1.22.3
pip==21.1.2
visitor==0.1.3
webcolors==1.11.1
wheel==0.36.2


